requirements.txt
Package                   Version
------------------------- ------------
altgraph                  0.17.4
CacheControl              0.13.1
cachetools                5.3.1
certifi                   2023.7.22
cffi                      1.15.1
charset-normalizer        3.2.0
contourpy                 1.1.1
cryptography              41.0.4
cycler                    0.11.0
DateTime                  5.2
firebase-admin            6.2.0
fonttools                 4.42.1
gcloud                    0.18.3
gevent                    23.9.1
google-api-core           2.12.0
google-api-python-client  2.101.0
google-auth               2.23.1
google-auth-httplib2      0.1.1
google-cloud-core         2.3.3
google-cloud-firestore    2.12.0
google-cloud-storage      2.11.0
google-crc32c             1.5.0
google-resumable-media    2.6.0
googleapis-common-protos  1.60.0
greenlet                  3.0.1
grpcio                    1.58.0
grpcio-status             1.58.0
httplib2                  0.22.0
idna                      3.4
importlib-metadata        6.8.0
importlib-resources       6.1.0
Jinja2                    3.1.2
jws                       0.1.3
kiwisolver                1.4.5
macholib                  1.16.3
MarkupSafe                2.1.3
matplotlib                3.8.0
msgpack                   1.0.6
numpy                     1.26.0
oauth2client              4.1.3
packaging                 23.1
pandas                    2.1.1
Pillow                    10.0.1
pip                       23.3.1
proto-plus                1.22.3
protobuf                  4.24.3
pyasn1                    0.5.0
pyasn1-modules            0.3.0
pycparser                 2.21
pycryptodome              3.19.0
pyinstaller               6.0.0
pyinstaller-hooks-contrib 2023.9
PyJWT                     2.8.0
pyparsing                 3.1.1
PyQt5                     5.15.10
PyQt5-Qt5                 5.15.11
PyQt5-sip                 12.13.0
pyqtgraph                 0.13.3
Pyrebase4                 4.7.1
pyserial                  3.5
PySide2                   5.15.2.1
python-dateutil           2.8.2
python-firebase           1.2
python-jwt                2.0.1
pytz                      2023.3.post1
qt-material               2.14
requests                  2.29.0
requests-toolbelt         0.10.1
rsa                       4.9
setuptools                58.1.0
shiboken2                 5.15.2.1
six                       1.16.0
tzdata                    2023.3
uritemplate               4.1.1
urllib3                   1.26.16
websocket                 0.2.1
zipp                      3.17.0
zope.event                5.0
zope.interface            6.0
